+++ src/util.h
@@ -4,6 +4,8 @@
 #include <iomanip>
 #include <iostream>
 #include <sstream>
+#include <fstream>
+#include <codecvt>
 
 namespace slip {
 
@@ -176,6 +178,11 @@ inline std::string escape_json(const std::string &s) {
         }
     }
     return o.str();
+}
+
+inline std::string to_utf8(const std::u16string &s) {
+  std::wstring_convert<std::codecvt_utf8<char16_t>, char16_t> conv;
+  return conv.to_bytes(s);
 }
 
 }